pyproject.toml
[tool.poetry]
name = "speechtotext"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
pyaudio = "^0.2.13"
google-cloud-speech = "^2.22.0"
google-cloud-texttospeech = "^2.14.2"
six = "^1.16.0"
speechrecognition = "^3.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
